Conflicts:
	src/main/java/net/therap/hyperbee/utils/constant/Messages.java
	src/main/java/net/therap/hyperbee/web/controller/ReservationController.java
	src/main/java/net/therap/hyperbee/web/helper/NoteHelper.java
	src/main/webapp/WEB-INF/decorators/topbar.jsp

+++ src/main/java/net/therap/hyperbee/web/controller/HiveController.java
@@ -11,6 +11,9 @@ import net.therap.hyperbee.web.helper.SessionHelper;
 import net.therap.hyperbee.web.validator.HiveValidator;
 import net.therap.hyperbee.web.validator.PostValidator;
 import net.therap.hyperbee.web.validator.UserIdInfoValidator;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.apache.logging.log4j.simple.SimpleLogger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -35,6 +38,8 @@ import static net.therap.hyperbee.utils.constant.Url.*;
 @Controller
 @RequestMapping("/user/hive")
 public class HiveController {
+
+    private static final Logger log = LogManager.getLogger(SimpleLogger.class);
 
     @Autowired
     private HiveService hiveService;
@@ -65,14 +70,28 @@ public class HiveController {
         binder.setValidator(postValidator);
     }
 
+    @InitBinder("userIdInfo")
+    private void initUserIdInfoBinder(WebDataBinder binder) {
+        binder.setValidator(userIdInfoValidator);
+    }
+
+    @InitBinder("hive")
+    private void hiveBinder(WebDataBinder binder) {
+        binder.setValidator(hiveValidator);
+    }
 
     @GetMapping
     public String viewHive(ModelMap model) {
         int userId = sessionHelper.getUserIdFromSession();
-        model.addAttribute("hiveList", hiveService.getHiveListByUserId(userId));
-        model.addAttribute("hive", new Hive());
+        model.addAttribute("hiveList", userService.findById(userId).getHiveList());
         model.addAttribute("userList", userService.findAll());
         model.addAttribute("userIdInfo", new UserIdInfo());
+
+        if (!model.containsAttribute("hive")) {
+            model.addAttribute("hive", new Hive());
+        }
+
+        log.debug("AuthUser ID: " + userId);
 
         return HIVE;
     }
@@ -86,37 +105,79 @@ public class HiveController {
         model.addAttribute("creator", userService.findById(hive.getCreatorId()));
         model.addAttribute("noticeList", hiveService.getLastFiveNotice(hive.getNoticeList()));
 
-        if(!model.containsAttribute("userIdInfo")){
+        if (!model.containsAttribute("userIdInfo")) {
             model.addAttribute("userIdInfo", new UserIdInfo());
         }
 
-        if(!model.containsAttribute("post")){
+        if (!model.containsAttribute("post")) {
             model.addAttribute("post", new Post());
         }
+
+        log.debug("Created Hive: " + hive.getName());
 
         return SHOW_HIVE;
     }
 
     @PostMapping(value = HIVE_ADD_USER_URL)
-    public String addUser(@ModelAttribute("userIdInfo") UserIdInfo userIdInfo, Model model, BindingResult result, RedirectAttributes redirectAttributes, @PathVariable("hiveId") int hiveId) {
+    public String addUser(@ModelAttribute("userIdInfo") UserIdInfo userIdInfo, Model model,
+                          BindingResult result, RedirectAttributes redirectAttributes,
+                          @PathVariable("hiveId") int hiveId) {
+
+        userIdInfoValidator.validate(userIdInfo, result);
+
+        if (result.hasErrors()) {
+            redirectAttributes.addFlashAttribute(BindingResult.MODEL_KEY_PREFIX + "userIdInfo", result);
+            redirectAttributes.addFlashAttribute("userIdInfo", userIdInfo);
+
+            return "redirect:" + HIVE_VIEW + hiveId;
+        }
 
         model.addAttribute("userInfoId", userIdInfo);
         hiveService.insertUsersToHive(hiveId, userIdInfo.getUserIdList());
+
+        log.debug("Member Added to hive : " + userIdInfo.getUserIdList());
 
         return "redirect:" + HIVE_VIEW + hiveId;
     }
 
     @PostMapping(value = HIVE_REMOVE_USER_URL)
-    public String RemoveUser(@ModelAttribute("userIdInfo") UserIdInfo userIdInfo, Model model,  BindingResult result, RedirectAttributes redirectAttributes, @PathVariable("hiveId") int hiveId) {
+    public String RemoveUser(@ModelAttribute("userIdInfo") UserIdInfo userIdInfo, Model model,
+                             BindingResult result, RedirectAttributes redirectAttributes,
+                             @PathVariable("hiveId") int hiveId) {
+
+        userIdInfoValidator.validate(userIdInfo, result);
+
+        if (result.hasErrors()) {
+            redirectAttributes.addFlashAttribute(BindingResult.MODEL_KEY_PREFIX + "userIdInfo", result);
+            redirectAttributes.addFlashAttribute("userIdInfo", userIdInfo);
+
+            return "redirect:" + HIVE_VIEW + hiveId;
+        }
 
         model.addAttribute("userInfoId", userIdInfo);
         hiveService.removeUsersFromHive(hiveId, userIdInfo.getUserIdList());
+
+        log.debug("Member Removed from hive : " + userIdInfo.getUserIdList());
 
         return "redirect:" + HIVE_VIEW + hiveId;
     }
 
     @PostMapping(value = HIVE_CREATE_URL)
-    public String saveHive(@ModelAttribute Hive hive, @RequestParam MultipartFile file, Model model) throws IOException {
+    public String saveHive(@Validated @ModelAttribute("hive") Hive hive, BindingResult result,
+                           RedirectAttributes redirectAttributes,
+                           @RequestParam MultipartFile file, Model model) throws IOException {
+
+        if (result.hasErrors() || file.getSize() == 0) {
+            redirectAttributes.addFlashAttribute(BindingResult.MODEL_KEY_PREFIX + "hive", result);
+            redirectAttributes.addFlashAttribute("hive", hive);
+
+            if (file.getSize() == 0) {
+                redirectAttributes.addFlashAttribute("fileError", "Please select picture");
+            }
+
+            return "redirect:" + HIVE_URL;
+        }
+
         model.addAttribute("hiveName", hive.getName());
         String filename = hive.getName().replaceAll(" ", "") + file.getOriginalFilename();
         hive.setImagePath(filename);
@@ -124,7 +185,6 @@ public class HiveController {
         hive.setCreatorId(userId);
         Hive newHive = hiveService.insertFirstUserToHive(hive, userId);
         hiveService.insertHive(newHive);
-        int hiveId = hiveService.getHiveIdByHiveName(newHive.getName());
 
         if (file.isEmpty()) {
         } else {
@@ -132,11 +192,15 @@ public class HiveController {
             model.addAttribute("message", imageUploader.createImage(filename, file));
         }
 
-        return "redirect:" + HIVE_VIEW + hiveId;
+        log.debug("AuthUser ID: " + userId);
+        log.debug("New Hive Name: " + hive.getName());
+
+        return "redirect:" + HIVE_VIEW + hiveService.getHiveByHiveName(newHive.getName()).getId();
     }
 
     @PostMapping(value = HIVE_ADD_POST_URL)
-    public String savePost(@Validated @ModelAttribute("post") Post post, BindingResult result, RedirectAttributes redirectAttributes, @PathVariable("hiveId") int hiveId) {
+    public String savePost(@Validated @ModelAttribute("post") Post post, BindingResult result,
+                           RedirectAttributes redirectAttributes, @PathVariable("hiveId") int hiveId) {
 
         if (result.hasErrors()) {
             redirectAttributes.addFlashAttribute(BindingResult.MODEL_KEY_PREFIX + "post", result);
@@ -147,6 +211,9 @@ public class HiveController {
 
         int userId = sessionHelper.getUserIdFromSession();
         postService.savePost(userId, hiveId, post);
+
+        log.debug("AuthUser ID: " + userId);
+        log.debug("Post: " + post.getDescription());
 
         return "redirect:" + HIVE_VIEW + hiveId;
     }
